refactor testing to show shared behavior
